requirements.txt
Flask==3.0.0
torch==2.6.0+cu124
torchvision==0.21.0+cu124
scikit-learn>=1.5.0
matplotlib>=3.9.0
seaborn>=0.13.0
Pillow==11.3.0
Werkzeug==3.0.1

